backend/django_main/requirements.txt
asgiref==3.5.2
Django==4.1
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.13.1
psycopg2==2.9.3
sqlparse==0.4.2
tzdata==2022.2